Own CNN

+++ src/main/java/com/strangerws/ssu/edu/textanalyzer/util/Dataset.java
@@ -8,7 +8,7 @@ import java.util.List;
 public class Dataset {
     private List<Entry> data;
 
-    public void setDataset(List<double[]> data, List<Double> result) {
+    public void setDataset(List<double[]> data, List<Integer> result) {
         if (data.size() != result.size()) {
             throw new RuntimeException("arrays length do not match each other");
         }
@@ -36,7 +36,7 @@ public class Dataset {
         Collections.shuffle(data);
     }
 
-    public void appendEntry(double[] data, double result) {
+    public void appendEntry(double[] data, int result) {
         this.data.add(new Entry(data, result));
     }
 
@@ -54,9 +54,9 @@ public class Dataset {
 
     public static class Entry {
         private double[] data;
-        private double result;
+        private int result;
 
-        public Entry(double[] data, double result) {
+        public Entry(double[] data, int result) {
             this.data = data;
             this.result = result;
         }
@@ -69,7 +69,7 @@ public class Dataset {
             this.data = data;
         }
 
-        public double getResult() {
+        public int getResult() {
             return result;
         }
